requirements.txt
numpy==1.21.6
Pillow==9.5.0
torch==1.13.1+cpu
torchvision==0.14.1+cpu
opencv-python==4.7.0.72
face-recognition==1.3.0
scikit-image==0.19.3
scipy==1.9.3
python-dateutil==2.8.2
tqdm==4.65.0
dlib==19.24.0
tensorflow-cpu==2.11.0  
kornia==0.6.9
albumentations==1.3.0